Youtube_Crawler/requirements.txt
requests
pymongo
python-dotenv
faktory
pyfaktory